Reverted EventHandler extraction (more tricky, to be done later). Extracted DiagnosticsDispatcher instead.
